// src/generated/BalanceResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BalanceResult = { isValid: boolean, remaining: number | null, used: number | null, total: number | null, unit: string | null, planName: string | null, invalidMessage: string | null, extra: unknown, };